style: Changes to structure - All features seperated from index
Index now just re-exports all relevant functions

BREAKING CHANGE: Restructured
