+++ src/GetOctetArray.ts
@@ -1,0 +1,3 @@
+export function getOctetArray (ipaddress: string | number): number[] {
+  return []
+}
