+++ src/IntToString.ts
@@ -1,0 +1,3 @@
+export function intToString (ipInt: number): string {
+  return ((ipInt >>> 24) + '.' + (ipInt >> 16 & 255) + '.' + (ipInt >> 8 & 255) + '.' + (ipInt & 255))
+}
