+++ src/index.ts
@@ -1,44 +1,5 @@
-const IPV4 = /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/
-const IPV6 = /^(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))$/
-
-export function isValidIp (ipaddress: any): Boolean {
-  return (isValidIpv4(ipaddress) || isValidIpv6(ipaddress))
-}
-
-export function isValidIpv4 (ipaddress: any): Boolean {
-  return IPV4.test(ipaddress)
-}
-
-export function isValidIpv6 (ipaddress: any): Boolean {
-  return IPV6.test(ipaddress)
-}
-
-export function stringToInt (ipaddress: String): Number {
-  return ipaddress.split('.').reduce(function (ipInt, octet) {
-    return (ipInt << 8) + parseInt(octet, 10)
-  } , 0) >>> 0
-}
-
-export function intToSring (ipInt: number): string {
-  return ((ipInt >>> 24) + '.' + (ipInt >> 16 & 255) + '.' + (ipInt >> 8 & 255) + '.' + (ipInt & 255))
-}
-
-export function isPrivate (ip: string | number): Boolean {
-  let ipaddress: string = ''
-  if (typeof ip === 'number') {
-    ipaddress = intToSring(ip)
-  } else {
-    ipaddress = ip
-  }
-  return /^(::f{4}:)?10\.([0-9]{1,3})\.([0-9]{1,3})\.([0-9]{1,3})$/i
-      .test(ipaddress) ||
-    /^(::f{4}:)?192\.168\.([0-9]{1,3})\.([0-9]{1,3})$/i.test(ipaddress) ||
-    /^(::f{4}:)?172\.(1[6-9]|2\d|30|31)\.([0-9]{1,3})\.([0-9]{1,3})$/i
-      .test(ipaddress) ||
-    /^(::f{4}:)?127\.([0-9]{1,3})\.([0-9]{1,3})\.([0-9]{1,3})$/i.test(ipaddress) ||
-    /^(::f{4}:)?169\.254\.([0-9]{1,3})\.([0-9]{1,3})$/i.test(ipaddress) ||
-    /^f[cd][0-9a-f]{2}:/i.test(ipaddress) ||
-    /^fe80:/i.test(ipaddress) ||
-    /^::1$/.test(ipaddress) ||
-    /^::$/.test(ipaddress)
-}
+export { isValidIp, isValidIpv4, isValidIpv6 } from './IsValidIp'
+export { stringToInt } from './StringToInt'
+export { intToString } from './IntToString'
+export { isPrivate } from './IsPrivate'
+export { getOctetArray } from './GetOctetArray'
